fix(lens): minor polish from the lens review
Separate reactive Dock row data from pull-request presentation.\n\n- lowy F6 — split RowPips along its independent change axes\n\nRaised by the lowy ∥ hickey lens review. Not pushed or merged.
