+++ packages/client/src/canvas/dock/Dock.tsx
@@ -117,7 +117,8 @@ import { type DockRowBucket, rowRecencyAt } from "./dockRowRanking";
 import type { DockGroup, DockTree } from "./dockTree";
 import { HiddenFooter } from "./HiddenFooter";
 import RecencyCell, { recencyMode } from "./RecencyCell";
-import { createDockRowData, PrPip } from "./RowPips";
+import { createDockRowData } from "./dockRowData";
+import { PrPip } from "./PrPip";
 import { rowSubline } from "./rowSubline";
 import { SubTerminalRow } from "./SubTerminalRow";
 import { useDockFocus } from "./useDockFocus";
