+++ packages/client/src/canvas/dock/DockList.tsx
@@ -38,7 +38,8 @@ import { type DockRowBucket, rowRecencyAt } from "./dockRowRanking";
 import type { DockGroup } from "./dockTree";
 import { HiddenFooter } from "./HiddenFooter";
 import RecencyCell, { recencyMode } from "./RecencyCell";
-import { createDockRowData, PrPip } from "./RowPips";
+import { createDockRowData } from "./dockRowData";
+import { PrPip } from "./PrPip";
 import { rowSubline } from "./rowSubline";
 import { SubTerminalRow } from "./SubTerminalRow";
 import { useDockOrder } from "./useDockOrder";
